feat: remove `client_actors` and `near_config` from indexer API (#14144)
This API stands in a way of making indexer testable because raw actors
are exposed. Instead we want to make indexer use senders instead to make
it compatible with testloop.

As an alternative now users can bootstrap `NearNode` in the client code
and then use `Indexer::from_near_node(..)`, see updated `CHANGELOG.md`
for more details.

We already expose a bunch of internals with `client_actors`, so using
`NearNode` isn't much worse. It also doesn't break existing client code
since the new way still gives access to all previously exposed data.

Note that the removed API was declared as unstable so it is not a
breaking change.

Part of #14109.