fix: take inflation rate from genesis for forknets (#14172)
There is some annoying part of code where for all reasonable protocol
versions inflation rate is always set to 5%, even though we have
`max_inflation_rate` parameter in genesis.
If forknet depends on the exact order of validators, non-zero rewards
may break it.

Based on #3294, `use_hardcoded_value` is just a hack to override
inflation. This hack was introduced because we didn't have epoch config
store yet. `max_inflation_rate` should be moved to epoch config,
especially if we decide to change inflation.

For now, let's change condition for `use_hardcoded_value` to checking
whether we are on mainnet or testnet. Due to limited replayability, it
is equivalent.
For forknets though, `max_inflation_rate` will be taken from their
genesis, as forknet engineers should expect.