fix(o11y): use tonic (grpc export) (#14183)
https://github.com/near/nearcore/pull/14163 bumps the version of
opentelemetry, however some of their API has changed.

As is node panics when exporting spans like:
```
   0: rust_begin_unwind
   1: core::panicking::panic_fmt
   2: tokio::task::spawn::spawn_inner::panic_cold_display
   3: tokio::time::sleep::Sleep::new_timeout
   4: reqwest::async_impl::client::Client::execute_request
   5: opentelemetry_http::reqwest::<impl opentelemetry_http::HttpClient for reqwest::async_impl::client::Client>::send_bytes::{{closure}}
   6: opentelemetry_sdk::trace::span_processor::BatchSpanProcessor::get_spans_and_export
note: Some details are omitted, run with `RUST_BACKTRACE=full` for a verbose backtrace.
```


[Next](https://github.com/open-telemetry/opentelemetry-rust/blob/main/docs/migration_0.28.md)
we learn that only `reqwest-blocking-client` can be used. However using
that we get some errors like:
```
2025-09-04T00:40:05.211682Z ERROR opentelemetry_sdk:  name="BatchSpanProcessor.ExportError" error="Operation failed: reqwest::Error { kind: Request, url: Url { scheme: \"http\", cannot_be_a_base: false, username: \"\", password: None, host: Some(Ipv4(10.164.0.4)), port: Some(4317), path: \"/\", query: None, fragment: None }, source: hyper_util::client::legacy::Error(SendRequest, hyper::Error(Parse(Version))) }"
```

This is because our metrics collector is using gRPC
[here](https://github.com/near/nearcore/blob/master/tracing/src/main.rs#L41)
("tonic" is the name of the library for gRPC). (Probably the dependency
for `tracing/` should be updated as well)

To get this to work I adjusted the feature flags and ran:
```
cargo update -p opentelemetry-otlp
cargo update -p hyper-util
```
to solve some dependency issues.

Not sure if somehow the setup prior to
https://github.com/near/nearcore/pull/14163 also enabled http/protobuf
export or only grpc/protobuf (http2).

Link to mocknet benchmark run:
https://github.com/Near-One/infra-ops/actions/runs/17451312198