feat(forknet): staking support (#14177)
For the tests where the number of validators increases, new joiners need
to send `stake` transaction , so that their proposal will be recorded in
protocol and they will join the validator set in the following epochs
([#releases/2.8 > 2.8 @
💬](https://near.zulipchat.com/#narrow/channel/522473-releases.2F2.2E8/topic/2.2E8/near/536851484))

The best way seems to be to set up near-cli and near-validators tool, as
they are designed to support it in human-friendly way.

It requires the following changes:
* Pass RPC node address to all the nodes, as they need to send stake tx
somewhere. I just use traffic' node address.
* Implement and upload `setup-near-cli.sh` and `send-stake-proposal.sh`
scripts to every node. Also, change the setup to `rsync` the whole
folder, as sending every script individually is too slow.
* setup-near-cli has minimal changes to support sending txs to forknet.
Unfortunately the main setup path is interactive, so I just copy
validator key to its internal storage.
* send-stake-proposal supports only integer amount of NEAR, so I need to
divide validator balance by `10**24`

Also a minor fix to `clear_scheduled_cmds` - we need to call
`reset-failed` too, otherwise the failed commands are not cleaned up.

I want to upload the 2.8 scenario in the following PR.

---------

Co-authored-by: Razvan Barbascu <razvan@nearone.org>
Co-authored-by: Razvan Barbascu <r.barbascu@gmail.com>