refactor: move `AccountContract` utilities off `TrieUpdate` (#14192)
`TrieUpdate` is a relatively weird place to put common logic such as
this, as it does not really have anything to do with account contracts
and its API surface should be somewhat agnostic of the data it is
working with.

But for the most part this is informed by the want to reuse this code
with the new `StateUpdate` and this logic living inside `TrieUpdate` was
bothering me quite some.

In fact this interface probably should take `&dyn TrieAccess` as an
argument or something of that sort, but this is not easily achievable
for `AccountContract::hash` due to `TrieAccess` not exposing accessors
for `OptimizedValueRef`s. Left the argument as `TrieUpdate` for now, but
in the future this will change (in a less involving way.)