refactor: move single-use utilities away from `TrieUpdate` (#14193)
These utility functions seem weird to have on a `TrieUpdate`. Usually I
would've considered adding an extension trait similar as for #14192, but
in this case these functions are only used once and adding an extension
trait seemed excessive. Especially when one of the uses is going away
shortly and the other is straightforward to replace with code that
already exists (that same `AccountContract` extension trait.)