feat(tracing): more spans (#14213)
More spans for viewing things in traviz.

Tx preparation:
* `prepare_transactions`

Chunk application stages:
* apply
* `TrieUpdate::finalize`
* process_transactions
* validate_apply_state_update

Also `apply_memtrie_changes`

I might be abusing the `block_production` tag a bit. I'm using it as a
marker for "I want this to be collected in benchmark spans". Maybe we
could consider `tag_traviz` for spans that should be exported for traviz
analysis?