feat(sync): avoid reapplying state parts (#14232)
Prior to this change it's possible for state parts to be re-applied if
state sync is interrupted at that stage.
This PR introduces a new DB column to track which state parts are
applied to avoid this issue.

Additionally, it will avoid deleting the flat storage if any parts are
applied and avoid recreating the flat storage if it was already
completed.

https://github.com/near/nearcore/issues/14032