cleanup(storage) remove mut from the storage opener  (#14276)
Removed mut from the config parameter to open_storage and instead assert
that config.archive has correct value. Mutating config is inconvenient
and can be surprising.

The first commit is the change, the rest is removing some other `mut`s
from usages.

Context: it annoyed me when working on the archival data loss recovery
because I couldn't store some references from inside the config since it
had to be mutable.