feat: spice: Implement separate witness for spice. (#14311)
We need the ability for witness to work when there are missing chunks.

Witness handling for spice doesn't have much common with existing
witness handling so it's much easier to introduce a completely different
struct compared to creating a new version. This also has an advantage of
allowing us to define and use types that best align with intuition for
spice witness handling.

Part of https://github.com/near/nearcore/issues/14300