feat: Action::DeterministicStateInit (#14307)
Add a new action `DeterministicStateInit` that creates an account with
verification that the id corresponds to a specific initial code and
state.

This is the main feature from NEP-616.
https://github.com/near/NEPs/pull/616

This PR also moves `GlobalContractIdentifier` from near-primitives to
near-primitives-core. This struct becomes part of how acocunt ids are
derived and therefore need to be accessible from the VM runtime. Smart
contracts will also need to access this struct type for verification of
predecessor ids.