create outcome for each transaction in the processed chunk (#14142)
Transactions that fail validation were dropped previously and now the
outcome will be registered in the `state_change` with the `Failed`
status.

- Extracted the single transaction processing into a function making
sure that every transaction generates an outcome (apart from
`RuntimeError` cases that result in chunk being dropped)
- added the Protocol feature to include every outcome into the
`state_change` instead of quietly dropping the transactions that failed
validation.
- added new error to signal the overflow when calculating the total gas
burnt by the chunk. This one I am not really sure about. This may not be
actionable by the party generating the transaction and the case may be
better handled by throwing `RuntimeError` and dropping the chunk as the
transaction gas usage alone has been already validated before.